moved operation componentes to specific directory
